좋아요,ㅈë¥ìì¦즐ê²ã겨찾기 완성

+++ src/main/java/com/example/ajouparking/repository/FavoriteRepository.java
@@ -9,10 +9,10 @@ public interface FavoriteRepository extends JpaRepository<Favorite,Long> {
 
     @Modifying
     @Query(value = "insert into favorite(user_id,parkinglot_id) values(:userId,:parkinglotId)",nativeQuery = true)
-    void addFavorite(int userId,Long parkinglotId);
+    void addFavorite(int userId,int parkinglotId);
 
     @Modifying
     @Query(value = "delete from favorite where user_id = :userId and parkinglot_id = :parkinglotId",nativeQuery=true)
-    void deleteFavorite(int userId, Long parkinglotId);
+    void deleteFavorite(int userId, int parkinglotId);
 
 }
